Additions to attributes dialog (close #1330)

+++ WinPort/src/sudo/sudo_common.cpp
@@ -76,11 +76,11 @@ namespace Sudo
 #if !defined(__APPLE__) && !defined(__FreeBSD__)
 # include <sys/ioctl.h>
 
-	int bugaware_ioctl_pint(int fd, unsigned long req, int *v)
+	int bugaware_ioctl_pint(int fd, unsigned long req, unsigned long *v)
 	{
 		// workaround for stupid FUSE/NTFS-3G bug with wrong FS_IOC_GETFLAGS' arg size,
 		// that modifies more than sizeof(int) near to arg ptr, corrupting memory
-		int placeholder[8] = {*v, 0};
+		unsigned long placeholder[8] = {*v, 0};
 		int out = ioctl(fd, req, &placeholder[0], NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL);
 		*v = placeholder[0];
 		return out;
